Grant leader lease to ranges; implement leader leases to be raft-agnostic
Added new Lease type to proto/data.proto. Leases contain an expiration
as a wall time in unix nanos as well as the node and store IDs of the
replica being granted the lease. Ranges keep track of the lease and can
now answer the question HasLeaderLease() for a range. Further, the last
set lease is used to construct proper NotLeaderError error returns.

On change in leader lease, only update timestamp cache by moving the low
water mark forward instead of clearing it completely. This change will also
need to rely on the last leader lease always being persisted.

Leases are raft-agnostic. They are a sequence of database time intervals
which are guaranteed not to overlap, for which a single replica in a raft
group has the leader lease. With the leader lease, a replica can submit
consistent read commands, write commands, and admin commands to raft, where
the command's timestamp is within the leader lease window. Note that this
is not the current system time, but the actual time of the command being
submitted (i.e. Request.Header.Timestamp).

Pre raft, when a command which requires the leader lease is received, the
lease holder is checked. If the lease is held by the replica and hasn't
expired (remember, according to the timestamp in the command's header),
the command can either be executed (consistent read and admin cases) or
submitted to raft (write case). If another replica currently holds the
lease and it hasn't expired (again, according to the timestamp in the
command's header), the a NotLeaderError is returned to the client, which
prompts a retry from the dist sender. If the lease currently isn't held,
the replica attempts to acquire it by submitting an InternalLeaderLease
command to raft. Because raft commands are executed by all replicas in
exact order, only one leader lease acquisition (the first after the prior
lease's expiration) will succeed.

Post raft, when a command is executed in Range.executeCmd, the lease is
verified.